port errors.rs test to context.rs
