+++ datafusion/core/src/execution/context.rs
@@ -2576,6 +2576,43 @@ mod tests {
         Ok(())
     }
 
+    #[tokio::test]
+    async fn unsupported_sql_returns_error() -> Result<()> {
+        let ctx = SessionContext::new();
+        ctx.register_table("test", test::table_with_sequence(1, 1).unwrap())
+            .unwrap();
+        let state = ctx.state();
+
+        // create view
+        let sql = "create view test_view as select * from test";
+        let plan = state.create_logical_plan(sql).await;
+        let physical_plan = state.create_physical_plan(&plan.unwrap()).await;
+        assert!(physical_plan.is_err());
+        assert_eq!(
+            format!("{}", physical_plan.unwrap_err()),
+            "This feature is not implemented: Unsupported logical plan: CreateView"
+        );
+        // // drop view
+        let sql = "drop view test_view";
+        let plan = state.create_logical_plan(sql).await;
+        let physical_plan = state.create_physical_plan(&plan.unwrap()).await;
+        assert!(physical_plan.is_err());
+        assert_eq!(
+            format!("{}", physical_plan.unwrap_err()),
+            "This feature is not implemented: Unsupported logical plan: DropView"
+        );
+        // // drop table
+        let sql = "drop table test";
+        let plan = state.create_logical_plan(sql).await;
+        let physical_plan = state.create_physical_plan(&plan.unwrap()).await;
+        assert!(physical_plan.is_err());
+        assert_eq!(
+            format!("{}", physical_plan.unwrap_err()),
+            "This feature is not implemented: Unsupported logical plan: DropTable"
+        );
+        Ok(())
+    }
+
     struct MyPhysicalPlanner {}
 
     #[async_trait]
