+++ datafusion/core/tests/sql/mod.rs
@@ -82,7 +82,6 @@ pub mod aggregates;
 #[cfg(feature = "avro")]
 pub mod avro;
 pub mod create_drop;
-pub mod errors;
 pub mod explain_analyze;
 pub mod expr;
 pub mod functions;
